feat(code): trying to fix bugs
