+++ header/GameMenu.h
@@ -1,20 +1,22 @@
 #ifndef GAMEMENU_H
 #define GAMEMENU_H
 
-#include "Timer.h"
-#include "Combo.h"
-#include "PlayerRecord.h"
-#include "Menu.h"
-#include "GameMenu.h"
+#include "../header/Timer.h"
+#include "../header/Combo.h"
+#include "../header/PlayerRecord.h"
+
+class Menu;
 
 class GameMenu {
 private:
     Timer time;
     Combo comboManager;
     PlayerRecord record;
-    Menu mainMenu;
+    Menu* mainMenu;
 
 public:
+    GameMenu();
+    ~GameMenu();
     void start();
     void trainingMode();
     void exit();
