+++ implementation/main.cpp
@@ -1,0 +1,16 @@
+#include "../header/GameMenu.h"
+#include "../header/Menu.h"
+#include "../header/PlayerRecord.h"
+#include <iostream>
+
+int main() {
+    PlayerRecord playerRecord;
+    GameMenu gameMenu;
+    Menu menu(playerRecord, gameMenu);
+
+    menu.showMenu();
+    //gameMenu.showGameMenu();
+
+    std::cout << "=== Exiting Program ===" << std::endl;
+    return 0;
+}
