--- src/implementation/main.cpp
@@ -1,10 +1,0 @@
-#include "GameMenu.h"
-#include <iostream>
-
-int main() {
-    GameMenu game;
-    game.showGameMenu();
-
-    std::cout << "=== Exiting Program ===" << std::endl;
-    return 0;
-}
